requirements.txt
jmetalpy~=1.7.0
pandas==2.2.2
numpy<=2.01
open3d==0.18.0
tqdm~=4.66.5
scipy~=1.14.1
scikit-learn==1.5.1
torch==2.0.0+cu118
torchaudio==2.0.0+cu118
torchvision==0.15.0+cu118
torchmetrics==1.4.2
pytorch-lightning==2.4.0
plotly~=5.23.0
matplotlib==3.9.2
laspy~=2.5.4
tensorboard<=2.17.1
plyfile~=1.1
numba~=0.60.0
networkx~=3.3
